Adjust Size of Multi-Select Labels
One of our bug reports was that clicking outside of a multi-select input would cause the same one (or a nearby one) to open - making it confusing and difficult to exit a multi-select box.

It turns out that this was because the labels had the class of 'col-md-2', making them (invisibly) very large. I've adjusted the markup to apply that 'col-md-2' class to a <div>, and not the label, leaving more 'dead space' that won't trigger the select boxes.

Note: I did not remove the ability to click the labels themselves, as that is an actual usability feature that I did not want to remove.