Move Column Classes Outside of Labels
Changed all applicable form components to use actual <div> elements with labels inside them, and not apply 'col-md-2' to labels. Adding that style to labels caused the labels to be very large, trapping clicks when we didn't want to. For accessibility reasons, we want to leave labels and click-able, we just don't want them to be that large.