Refactor Javascript for DropDown Lists
Refactored the javascript used for drop-down lists so that it will work with one or two drop-down lists. The old javascript added the word 'undefined' to the search string when one of the drop-down lists was missing. This code should work in both cases, which will hopefully allow us to extract this code into a stand-alone script to keep things DRY.